feat: Add payment system

+++ WrocRide/Models/DriverDto.cs
@@ -7,7 +7,7 @@ namespace WrocRide.Models
         public string Name { get; set; }
         public string Surename { get; set; }
         public float? Rating { get; set; }
-        public float Pricing { get; set; }
+        public decimal Pricing { get; set; }
         public DriverStatus DriverStatus { get; set; }
     }
 }
